pre-open-scanner-main/backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
websockets==12.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0
pyyaml==6.0.1
pandas==2.1.4
redis==5.0.1
python-dotenv==1.0.0
requests==2.31.0
openpyxl==3.1.2
